OneTakeAnalytics/pyproject.toml
[project]
name = "onetake-analytics"
version = "0.1.0"
description = "OneTake Analytics and Recommendations (gRPC services, ClickHouse)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "clickhouse-driver>=0.2.6",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "libs/onetake_common",
    "libs/onetake_proto",
    "libs/onetake_analytics_model",
    "services/analytics_ingest",
    "services/reco_service",
]

[tool.hatch.build.targets.sdist]
include = [
    "libs",
    "services",
    "contracts",
]
